Revert "Address clippy complaints in src/vm/"
This reverts commit 264831e1ba86e16bf480237a1c76182c23974866.

+++ clarity/src/vm/representations.rs
@@ -119,8 +119,8 @@ impl StacksMessageCodec for ClarityName {
 
 impl StacksMessageCodec for ContractName {
     fn consensus_serialize<W: Write>(&self, fd: &mut W) -> Result<(), codec_error> {
-        if self.as_bytes().len() < CONTRACT_MIN_NAME_LENGTH
-            || self.as_bytes().len() > CONTRACT_MAX_NAME_LENGTH
+        if self.as_bytes().len() < CONTRACT_MIN_NAME_LENGTH as usize
+            || self.as_bytes().len() > CONTRACT_MAX_NAME_LENGTH as usize
         {
             return Err(codec_error::SerializeError(format!(
                 "Failed to serialize contract name: too short or too long: {}",
